+++ client/src/components/ProfileDropdown/ProfileDropdown.scss
@@ -5,23 +5,18 @@
   z-index: 90;
   width: 100%;
   height: 100%;
-
-  backdrop-filter: blur(4px);
-  // opacity: 0;
-  // transform: scale(0);
-  transition: opacity 400ms, transform 400ms;
+  transition: opacity 300ms, backdrop-filter 300ms;
 }
 
 .profile-dropdown-bg-display {
   opacity: 1;
-  transform: scale(1);
-  transition: opacity 400ms, transform 400ms;
+  backdrop-filter: blur(4px);
 }
 
 .profile-dropdown-bg-hide {
+  pointer-events: none;
   opacity: 0;
-  transform: scale(0);
-  transition: opacity 400ms, transform 400ms;
+  backdrop-filter: blur(0px);
 }
 
 .profile-dropdown-container {
